requirements.txt
tornado==6.0.2
mysql-connector-python==8.0.16
sqlalchemy==1.3.3
psutil==5.6.2
pyecharts==1.1.0
pyecharts-snapshot==0.2.0
